requirements.txt
django==3.1.5
feedparser==6.0.2
psycopg2-binary==2.8.6
django-crispy-forms==1.10.0
django-markdown-deux==1.0.5
